Add comments to code and last scan date to shared preferences

+++ app/src/main/java/horvatApps/ImageScan/ui/MainActivity.java
@@ -38,19 +38,15 @@ import horvatApps.ImageScan.db.models.Mapper;
 
 public class MainActivity extends AppCompatActivity {
 
-    private Toolbar toolbar;
-    private AppBarLayout appBarLayout;
-    private RecyclerView recyclerView;
     private RecyclerViewAdapter recyclerViewAdapter;
 
-    private ArrayList<ImageDetail> imageList;
-    private ArrayList<ImageDetail> searchedImages;
     private ArrayList<String> imageFolders;
     private ArrayList<ImageDetail> imagesFromDB = new ArrayList<ImageDetail>();
 
     private MainViewModel mainViewModel;
 
     @Override
+    //initialises the viewModel, UI elements and observers for live data
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
@@ -61,10 +57,11 @@ public class MainActivity extends AppCompatActivity {
 
     }
 
+    //checks if any scans were already performed
     @Override
     protected void onResume() {
         super.onResume();
-        permissionHandle();
+        handleSharedPref();
     }
 
     /*
@@ -75,15 +72,18 @@ public class MainActivity extends AppCompatActivity {
     private SearchView searchView;
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
+        //create the toolbar menu
         MenuInflater inflater = getMenuInflater();
         inflater.inflate(R.menu.toolbar_menu, menu);
 
+        //create the search bar in the toolbar menu
         MenuItem searchItem = menu.findItem(R.id.toolbar_search);
         searchView = (SearchView) searchItem.getActionView();
         searchView.setIconifiedByDefault(false);
 
         recyclerViewAdapter.notifyDataSetChanged();
 
+        //handle searching in the search bar
         searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
             @Override
             public boolean onQueryTextSubmit(String query) {
@@ -99,6 +99,7 @@ public class MainActivity extends AppCompatActivity {
         return true;
     }
 
+    //handles clicking on icons from the toolbar menu
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()){
@@ -125,12 +126,12 @@ public class MainActivity extends AppCompatActivity {
 
 
     public void initUiElements() {
-        appBarLayout = findViewById(R.id.app_bar_layout);
-        toolbar = findViewById(R.id.toolbar);
+        AppBarLayout appBarLayout = findViewById(R.id.app_bar_layout);
+        Toolbar toolbar = findViewById(R.id.toolbar);
         toolbar.setTitle(R.string.app_name);
         setSupportActionBar(toolbar);
 
-        recyclerView = findViewById(R.id.recyclerImages);
+        RecyclerView recyclerView = findViewById(R.id.recyclerImages);
         recyclerView.setLayoutManager(new GridLayoutManager(this, 3));
         recyclerViewAdapter = new RecyclerViewAdapter(this, imagesFromDB);
         recyclerView.setAdapter(recyclerViewAdapter);
@@ -142,7 +143,9 @@ public class MainActivity extends AppCompatActivity {
     ----------------------------------------------------------------------------------------------------------
      */
 
+    //initialises observers for live data
     public void observerSetup() {
+        //handle all data live updates in recycler view if not using the search bar
         mainViewModel.getAllImages().observe(this, new Observer<List<ImageEntityDB>>() {
 
             @Override
@@ -162,7 +165,9 @@ public class MainActivity extends AppCompatActivity {
                             imagesFromDB.add(Mapper.imageEntityToImageDetail(imageEntityDB));
                         }
 
+                        //clears deleted images from database
                         clearGoneImages();
+
                         recyclerViewAdapter.notifyDataSetChanged();
                     } catch (Exception e) {
                         e.printStackTrace();
@@ -170,6 +175,7 @@ public class MainActivity extends AppCompatActivity {
             }
         });
 
+        //handle search bar live updates in recycler view
         mainViewModel.getSearchedImages().observe(this, new Observer<List<ImageEntityDB>>() {
             @Override
             public void onChanged(List<ImageEntityDB> imageEntityDBS) {
@@ -190,28 +196,34 @@ public class MainActivity extends AppCompatActivity {
     }
 
     /*
-    GET ALL SCANNED IMAGES
+    HANDLES REMOVED IMAGES FROM DEVICE
     ----------------------------------------------------------------------------------------------------------
      */
 
+    //removes all images stored in the database that are not on device anymore
     private void clearGoneImages() {
-        ArrayList<String> allImageUris = getAllImageUrisOnDevice();
-        ArrayList<String> imagesToDelete = new ArrayList<String>();
+        //check for files acces permission
+        if (ContextCompat.checkSelfPermission(this, Manifest.permission.READ_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
 
-        for(ImageDetail imageDetail : imagesFromDB){
-            if(!allImageUris.contains(imageDetail.getUri().toString()))
-                imagesToDelete.add(imageDetail.getUri().toString());
+            ArrayList<String> allImageUris = getAllImageUrisOnDevice();
+            ArrayList<String> imagesToDelete = new ArrayList<String>();
+
+            for (ImageDetail imageDetail : imagesFromDB) {
+                if (!allImageUris.contains(imageDetail.getUri().toString()))
+                    imagesToDelete.add(imageDetail.getUri().toString());
+            }
+
+            if (imagesToDelete.size() > 0)
+                mainViewModel.deleteImages(imagesToDelete);
         }
-
-        if(imagesToDelete.size() > 0)
-            mainViewModel.deleteImages(imagesToDelete);
     }
 
+    //fetches uri of all images on device
     private ArrayList<String> getAllImageUrisOnDevice() {
         ArrayList<String> imageUris = new ArrayList<String>();
 
+        //sql select
         String[] projection = {MediaStore.Images.ImageColumns._ID};
-
 
         try (Cursor cursor = this.getContentResolver().query(android.provider.MediaStore.Images.Media.EXTERNAL_CONTENT_URI,
                 projection, null, null, null)) {
@@ -241,67 +253,14 @@ public class MainActivity extends AppCompatActivity {
 
     public void handleSharedPref() {
         SharedPreferences sharedPref = getApplicationContext().getSharedPreferences("ImageScanPref", 0);
-        boolean scanned = sharedPref.getBoolean("Scanned", false);
-        if (!scanned) {
-            /*Intent intent = new Intent(this, SettingsActivity.class);
-            startActivity(intent);*/
-        } else {
+        String lastScanTime = sharedPref.getString("LastScan", "never");
 
-            HashSet<String> selectedFolders = new HashSet<>();
-
-            imageFolders = new ArrayList<String>(sharedPref.getStringSet("Folders", new HashSet<String>()));
-
-            //readImages();
-        }
-    }
-
-
-    /*
-    PERMISSIONS
-    ----------------------------------------------------------------------------------------------------------
-     */
-
-    public void permissionHandle() {
-        if (ContextCompat.checkSelfPermission(this, Manifest.permission.READ_EXTERNAL_STORAGE) ==
-                PackageManager.PERMISSION_GRANTED) {
-            // You can use the API that requires the permission.
-            Toast.makeText(this, "permission granted", Toast.LENGTH_LONG).show();
-            handleSharedPref();
-
-        } else if (shouldShowRequestPermissionRationale(Manifest.permission.READ_EXTERNAL_STORAGE)) {
-            // In an educational UI, explain to the user why your app requires this
-            // permission for a specific feature to behave as expected. In this UI,
-            // include a "cancel" or "no thanks" button that allows the user to
-            // continue using your app without granting the permission.
-            Toast.makeText(this, "rationale", Toast.LENGTH_LONG).show();
-
-        } else {
-            String[] premissions = {Manifest.permission.READ_EXTERNAL_STORAGE};
-            requestPermissions(premissions, 420);
-        }
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-
-        switch (requestCode) {
-            case 420:
-                // If request is cancelled, the result arrays are empty.
-                if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                    // Permission is granted. Continue the action or workflow
-                    // in your app.
-                    Toast.makeText(this, "111", Toast.LENGTH_LONG).show();
-                    handleSharedPref();
-                } else {
-                    // Explain to the user that the feature is unavailable because
-                    // the features requires a permission that the user has denied.
-                    // At the same time, respect the user's decision. Don't link to
-                    // system settings in an effort to convince the user to change
-                    // their decision.
-                    Toast.makeText(this, "222", Toast.LENGTH_LONG).show();
-                }
-                return;
+        //checks shared prefferences for the date of last scan
+        assert lastScanTime != null;
+        if (lastScanTime.equals("never")) {
+            //if scan not preformed yet redirect to scan activity
+            Intent intent = new Intent(this, ScanActivity.class);
+            startActivity(intent);
         }
     }
